requirements.txt
biopython==1.81
numpy==1.24.2